delete redundant files in data_utils
