--- core/data_utils/load_pubmed_lp.py
@@ -1,61 +1,0 @@
-import os, sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-import numpy as np
-# adapted from https://github.com/jcatw/scnn
-import torch
-from torch_geometric.datasets import Planetoid
-import torch_geometric.transforms as T
-from sklearn.preprocessing import normalize
-from torch_geometric.transforms import RandomLinkSplit
-from data_utils.load_pubmed import parse_pubmed
-from torch_geometric.data import Data
-
-def get_pubmed_casestudy(args):
-    corrected = False
-    undirected = args.data.undirected
-    include_negatives = args.data.include_negatives
-    val_pct = args.data.val_pct
-    test_pct = args.data.test_pct
-    split_labels = args.data.split_labels
-    
-    _, data_X, data_Y, data_pubid, data_edges = parse_pubmed()
-    data_X = normalize(data_X, norm="l1")
-
-    # load data
-    dataset = Planetoid('./generated_dataset', 'PubMed', transform=T.NormalizeFeatures())
-    data = dataset[0]
-
-    # replace dataset matrices with the PubMed-Diabetes data, for which we have the original pubmed IDs
-    x = torch.tensor(data_X)
-    edge_index = torch.tensor(data_edges)
-    y = torch.tensor(data_Y)
-    num_nodes = data.num_nodes
-    
-    # split data
-    if corrected:
-        is_mistake = np.loadtxt(
-            'pubmed_casestudy/pubmed_mistake.txt', dtype='bool')
-        train_id = [i for i in train_id if not is_mistake[i]]
-        val_id = [i for i in val_id if not is_mistake[i]]
-        test_id = [i for i in test_id if not is_mistake[i]]
-
-
-    data = Data(x=x,
-        edge_index=edge_index,
-        y=y,
-        num_nodes=num_nodes,
-        node_attrs=x, 
-        edge_attrs = None, 
-        graph_attrs = None
-    )        
-
-    undirected = data.is_undirected()
-
-    transform = RandomLinkSplit(is_undirected=undirected, num_val=val_pct, num_test=test_pct,
-                                add_negative_train_samples=include_negatives, split_labels=split_labels)
-    train_data, val_data, test_data = transform(dataset._data)
-    splits = {'train': train_data, 'valid': val_data, 'test': test_data}
-    
-    dataset._data = data
-    
-    return dataset, data_pubid, splits
